Extract session holder 1235
